refactor(climb_tower_agent): climb_tower模组化
