+++ agent/custom/action/__init__.py
@@ -1,4 +1,6 @@
-from .climb_tower import *
+from .climb_tower_shop import *
+from .climb_tower_preparation import *
+from .climb_tower_loop import *
 from .invite import *
 from .fight import *
